Create a plugin-commons module
