+++ plugin-commons/src/main/resources/plugins_min_versions.txt
