+++ pom.xml
@@ -31,6 +31,7 @@
     <module>server-api</module>
     <module>telemetry</module>
     <module>issue-tracking</module>
+    <module>plugin-commons</module>
   </modules>
 
   <properties>
